requirements.txt
pytest == 4.0.0
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
lxml-4.3.3-cp37-cp37m-win32.whl
